Make sure we scope systemctl service changes to install only.

+++ supervisor-status/supervisor-status.spec
@@ -2,7 +2,7 @@
 
 Name:       supervisor-status
 Version:    1.0.0
-Release:    2%{?dist}
+Release:    3%{?dist}
 Summary:    Service that reports current supervisor status as JSON.
 License:    MIT
 Group:      System Environment/Libraries
@@ -44,14 +44,18 @@ rm -rf %{buildroot}
 %attr(0744,root,root)/usr/lib/systemd/system/supervisor-status.socket
 
 %post
-systemctl enable supervisor-status.socket
-systemctl start supervisor-status.socket
+if [ $1 -eq 1 ] ; then
+  systemctl enable supervisor-status.socket
+  systemctl start supervisor-status.socket
+fi
 
 %preun
-systemctl stop supervisor-status.service
-systemctl disable supervisor-status.service
-systemctl stop supervisor-status.socket
-systemctl disable supervisor-status.socket
+if [ $1 -eq 1 ] ; then
+  systemctl stop supervisor-status.service
+  systemctl disable supervisor-status.service
+  systemctl stop supervisor-status.socket
+  systemctl disable supervisor-status.socket
+fi
 
 %changelog
 * Wed Jun 14 2017 Joe Grund
